euler methods

+++ ntsa/tools/data_viz.py
@@ -8,7 +8,7 @@ class DataViz:
 
     def plot_data(self, data, x="t", y="y", label=None):
         ax = plt.gca()
-        df = pd.DataFrame(data)
+        df = pd.DataFrame(data).astype(float)
         df.plot(kind="line", x=x, y=y, ax=ax, label=label)
 
     def show(self):
